requirements.txt
altgraph==0.17.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
packaging==25.0
pycparser==2.23
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
setuptools==80.9.0
shiboken6==6.9.2
